introduce per-model uniform struct
Right now the per-model uniform only carries a per-model local to world
space matrix transform. In the future it will probably have a lot more
stuff such as a list of lights for rendering.

I renamed the previous 'per-model' struct to `per-mesh` which better
reflects that its texture view and sampler is used for one or more
submeshes per model.

Finally I added a whole bunch of TOOD notes for the future when I want
to render multiple models that potentially have multiple meshes.