depth buffer visualization
Enhances the depth pass with the ability to render the depth buffer to a
full screen quad for visualization. You can toggle the depth buffer by
pressing 'z' while running.

The depth buffer pass is reasonably self contained so it should be
straight forward to drop it into another project.