refactor per-* uniform buffers to trait+struct
A significant majority of the `per-*` uniform structs share the same
code. I unified the `per-model` and `per-frame` structs with a new trait
representing uniform buffers, as well as a helper struct to simplify
the copying of values.

I'm not 100% sure this is the correct approach yet but I want to
experiment with ways to simplify mirroring uniform values in the
renderer and in the shader. I'm planning on adding lighting soon, plus
other potential debug shaders need uniform values too.

I'll take another pass at this once I've thought through how to support
texture views and samplers (hopefully with these types rather than more
abstraction).