rudimentary support for materials and lights
This change adds a simple materials struct that can pass constant values
for blinn-phong material properties to submeshes. It also introduces a
light struct that has a position, color and a diffuse and specular
contribution modifier property.

The demo has been updated to have a light that orbits the cube along
with a color that slowly varies over time.