remove custom Texture type
This refactor change removes the custom Texture type that I introduced
close to the start of the project. It served its purpose well at the
start but has since become restrictive as I add multiple texture maps to
the standard material type. The primary problem was it bound a sampler
and view to every texture that was created when I wanted to share
samplers between multiple textures in the same material.

After removing the custom type I switched the functions to returning
simple `wgpu::Texture` objects. The view can be easily derived from it.
Samplers can be manually created now since they have no direct
relationship to a texture or view.