directional lights
Adds a directional light to the standard shader which can be set in the
`per_frame` uniform buffer. The implementation of directional lights is
very similiar to the pseudo point light implementation previously but
with direction instead of position.

I also took some time to refactor the standard shader code to use
functions for lighting calculations for code re-use. Right now the
paramters lists are gigantic but I added a TODO for later to experiment
with structs to make things more self-documenting. For the time being I
added a ton of comments to maybe make it easier to follow.