refactor to use instancing in light debug pass
I switched up the light debug pass to support an arbitrary number of
lamp objects for a future demo. While doing this I tried to switch the
pass to using instance rendering and took a crack at making the code
similiar to `GenericUniformBuffer`.

Honestly I'm not super happy with the initial implementation as it leaks its
implementation details all over the place. I'll keep with the
philosophy of this project ("move fast and don't prematurely make things
reusable until you understand how it be reused") and leave it alone
until the future when I have a better handle on instancing.