fix windows build break
This was an odd compile error that showed up when I switched from my
Macbook to my Windows desktop. It looks like the `glam` crate has
different implementations for `Vec4`? On Windows it doesn't allow direct
field instantiation but it works fine on Mac. Weird!

Also addressed a few minor linter issues that again seemed to only pop
up on Windows. This could also be triggered by an older version of Rust
on the Mac