refactor out demo scene code
This is a **very** simplistic refactor that tries to move a bunch of the
gameplay-esque demo code out of the renderer and into a separate class.
This new gameplay trait `GameApp` is still deeply intertwined with the
renderer (it alters its properties!) and is not a good example of
separating the renderer out from the game.

I did this as the simplest possible cut so that I could continue
iterating on renderer features. The renderer attempts to be somewhat
scene agnostic (although with hardcoded features), and `GameApp`
implementations contain most of the demo code. I'll revisit a lot of
this in the future once I'm happier with the features I've jammed into
the renderer.

Again I'm trying to sprint on trying to try out graphical features in
this project. Once I'm happy with a bunch of it I'll understand better
how to refactor everything into clean systems. Otherwise I risk making
incorrect decisions before I have sufficient information.