introduce scene organization
I experimented with moving a bunch of "scene" properties from the
renderer into a new scene object that can be passed between the gameplay
code and the renderer. It's important to note this scene isn't a scene
graph - it can be easily recreated every frame if needed. The scene is
meant to be a collection of all the per-frame values that a renderer
would need to draw the frame.

This change also tried something new and I think it worked out well. It
moves the storage of model shader values back to the renderer and
instead hands out `SlotMap` keys to the shader values. I did this to
keep the shader values (which only the renderer cares about) out of the
model and hence out of the scene. There's more work to be done, but I
want to get to a place where `Model` and `Scene` don't have any special
renderer knowledge.